Force dtype to int32

+++ yasa/tests/test_hypno.py
@@ -77,13 +77,14 @@ class TestHypno(unittest.TestCase):
         expected = pd.DataFrame({
             'values': [0, 1, 0, 1, 0],
             'start': [0, 11, 14, 16, 25],
-            'length': [11, 3, 2, 9, 2]})
+            'length': [11, 3, 2, 9, 2]}, dtype=np.int32)
 
         assert_frame_equal(hfp(x, sf_hypno=1 / 60, threshold="0min"), expected)
         assert_frame_equal(hfp(x, sf_hypno=1, threshold="0min"), expected)
 
         # 1b. With thresholding
-        expected = pd.DataFrame({'values': [0, 1], 'start': [0, 16], 'length': [11, 9]})
+        expected = pd.DataFrame(
+            {'values': [0, 1], 'start': [0, 16], 'length': [11, 9]}, dtype=np.int32)
         assert_frame_equal(hfp(x, sf_hypno=1 / 60, threshold="5min"), expected)
         assert hfp(x, sf_hypno=1, threshold="5min").size == 0
 
@@ -91,7 +92,7 @@ class TestHypno(unittest.TestCase):
         expected = pd.DataFrame({
             'values': [0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 1, 0],
             'start': [0, 2, 4, 6, 8, 11, 14, 16, 18, 20, 22, 25],
-            'length': [2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2]})
+            'length': [2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2]}, dtype=np.int32)
         assert_frame_equal(hfp(x, sf_hypno=1 / 60, threshold="2min", equal_length=True), expected)
 
         # TEST 2: MULTI-CLASS VECTOR
@@ -100,6 +101,5 @@ class TestHypno(unittest.TestCase):
         expected = pd.DataFrame({
             'values': [0, 1, 2, 0, 1, 0, 1],
             'start': [0, 4, 5, 11, 14, 15, 16],
-            'length': [4, 1, 6, 3, 1, 1, 1]})
-
+            'length': [4, 1, 6, 3, 1, 1, 1]}, dtype=np.int32)
         assert_frame_equal(hfp(x, sf_hypno=1 / 60, threshold="0min"), expected)
